Enable image completion on all platforms
