+++ cmd/nerdctl/container/container_run.go
@@ -465,3 +465,10 @@ func runAction(cmd *cobra.Command, args []string) error {
 	}
 	return nil
 }
+
+func runShellComplete(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+	if len(args) == 0 {
+		return completion.ImageNames(cmd)
+	}
+	return nil, cobra.ShellCompDirectiveNoFileComp
+}
